RangedBird
Made RangedBird, so far it is a stork. It flies in the top of the screen, in random heights, and when it is near the player it drops a bird poop. Sometimes is doesn't hit, but most of the time, it does. This can be changed by making the random float a higher margin.