feathers and fixed destroying of poop
I have made a simple particle system for some white feathers. Not the prettiest, but it might do for now.
I have also fixed the issue where the poop that a ranged bird just dropped, got destroyed if the bird that dropped it got hit by stone. This is done a way, where it just makes the bird invisble, and destroyed after 3 seconds. Then if it has just dropped a poop, it will have time to either hit the player or the ground.