Copies of gamescene
I havve made 4 copies of the gamescene I could get to work. These gamescenes do not update the score, and do not have the health. Maybe it wasn't comittet, the SendMessage("updateScore") came back with an error, that it could not be found.
Now we work on our own gamescenes when testing stuff out, and creating new prefabs and such, and when we are totally sure works we can make it in the original gamescene.
We DO NOT make anything in the original gamescene without first telling everyone in the group. And we tell everyone when we are finished making the changes in the original, so another one can start making his changes, if we are all working on the game at the same time.
This way, we will maybe not get all these merging errors.