Made touchInput script
I have made touchInput scriptet. So far it makes the player run like the keyboard script, though it takes the ruinningSpeed variable from the playerStats script, that's more logical.
And on touchEnded it throws a stone at the direction you touched. Now we gotta make it check if you touch a gameobject, because then we can make a couple buttons to make him jump and protect himself! These buttons are just gonna be children of the camera object, so it moves with that. This way, we dont need a script to make them stay where they are supposed to stay.