Started on Tier/Stage
I have started working on the tiers/stages. It now changes the birds health and speed, when the tier gets higher, the health and speed gets higher on the meleeBirds.
I have also added a lot of different bird sprites, it is the same artstyle as the yellow bird.

I ahve also added a lot of other graphics, so we can use that later on. Like tree, signs, different health sprites and much much more.