First mini boss + A lot more changes
I made the first Mini Boss. There is still some tweaking to do with it's script, so it looks more smooth.
This type of mini boss, is a green dragon that flies above the player from side to side, like the ranged birds. But it attacks with melee like the melee birds. So every time it flies by the player, it will fly down and hit him for X amount of damage and then it will fly up into the air again and start all over. It will do this until it is killed. It will spawn when there is an increase in Stage/Tier/Level, and when there is a boss fight going on, there will not spawn regular birds.

I fixed the point at where the birds change direction, it will now change direction when it hits the screen border, no matter how long or narrow it is.

I have made a second weapon type, the axe. With this I have also made an weaponUpgrade resource, when it is collected, the players weapon type will change. If there are no more greater weapon types, he will just keep the one he already had.
I added some rotation to the weapon ammo, when there comes a weapon type like rifle, we need to change it a bit, so the bullets don't rotate. But I can do that, and should be pretty easy.