Endless Background + small btn changes
I have made the background endless now. It is not paralax yet, but I have an idea about how to do it now.
I also made a small change in the button script. They now position themselves in the bottom left corner of the screen. So no matter what size/resolution/ratio the screen is, they will always be positioned in the bottom left corne with an offset.

Plus, We need to change working enviroment to Android. And then choose a resolution to go from. There is an Asset in the asset store, called "Auto Letterbox" that will help out tremendously with this. If it is a screen not fit for the game, it will make it fit with black bars on top and bottom or left and right, so it won't look stretched or loose some of the sides.