Paralax Background
Added paralax background, so the mountains move a bit quicker than the sky. It is still endless.

I have also created 2 more bird prefabs, and added them in the birdspawner. So when a tier increase it will start spawning another bird. But there is only 3 different meleeBirds. the yellow, a brown and a blue bird. They have different damage, speed and stuf like that.
The blue bird will for example need 2 axes to the face, before it dies because it has 30 health.